initial volume list generator

+++ app.py
@@ -1,35 +1,56 @@
-import setup
-from lib import apps_oper
+from helpers.logger import Colors
 from lib import services_oper
+from lib import volumes_oper
 from lib import file_oper
+from lib import apps_oper
+import setup
 
 
 def main():
-    content = ""
+    if not setup.GENERATE_VOLUME_FILE and not setup.GENERATE_PORT_FILE:
+        print(Colors.RED + "No files to generate")
+        exit(1)
+    if setup.GENERATE_VOLUME_FILE:
+        all_volumes = []
+        volume_file_content = ""
+        volume_file_content += setup.VOLUME_LIST_INTRO
+        file_oper.delete_file(setup.VOLUME_LIST_FILE)
+
+    if setup.GENERATE_PORT_FILE:
+        all_ports = []
+        port_file_content = ""
+        port_file_content += setup.PORT_LIST_INTRO
+        file_oper.delete_file(setup.PORT_LIST_FILE)
+
     ordered_trains = apps_oper.get_ordered_trains(apps_oper.get_trains())
-    file_oper.delete_file(setup.PORT_LIST_FILE)
-    # file_oper.delete_file(setup.VOLUME_LIST_FILE)
-    # file_oper.add_text_to_file(setup.VOLUME_LIST_FILE, setup.VOLUME_LIST_INTRO)
-    all_ports = []
-    content += setup.PORT_LIST_INTRO
     for train in ordered_trains:
         apps = apps_oper.get_apps(train)
-        # Add the list that is returned to the services_list,
-        # can't append is we will end up with a list of lists
-        raw_services_list = (services_oper.get_raw_services_list(apps))
-        all_ports += services_oper.get_processed_services_list(
-            raw_services_list, train.stem)
-        # Later we can add a volumes_list etc
-        # file_oper.create_volume_list_file(volume_list, train.stem)
+        if setup.GENERATE_PORT_FILE:
+            # Add the list that is returned to the raw_services_list,
+            # can't append or we will end up with a list of lists
+            raw_services_list = services_oper.get_raw_services_list(apps)
+            all_ports += services_oper.get_processed_services_list(
+                raw_services_list, train.stem)
+        if setup.GENERATE_VOLUME_FILE:
+            raw_volumes_list = volumes_oper.get_raw_volumes_list(apps)
+            all_volumes += volumes_oper.get_processed_volumes_list(
+                raw_volumes_list, train.stem)
 
     services_oper.append_conflicts_to(all_ports)
     for train in ordered_trains:
-        content += services_oper.create_port_list_content(
-            all_ports, train.stem)
+        if setup.GENERATE_PORT_FILE:
+            port_file_content += services_oper.create_port_list_content(
+                all_ports, train.stem)
+        if setup.GENERATE_VOLUME_FILE:
+            volume_file_content += volumes_oper.create_volume_list_content(
+                all_volumes, train.stem)
 
-    content += setup.PORT_LIST_OUTRO
-    file_oper.add_text_to_file(setup.PORT_LIST_FILE, content)
-    # file_oper.add_text_to_file(setup.VOLUME_LIST_FILE, setup.VOLUME_LIST_OUTRO)
+    if setup.GENERATE_PORT_FILE:
+        port_file_content += setup.PORT_LIST_OUTRO
+        file_oper.add_text_to_file(setup.PORT_LIST_FILE, port_file_content)
+    if setup.GENERATE_VOLUME_FILE:
+        volume_file_content += setup.VOLUME_LIST_OUTRO
+        file_oper.add_text_to_file(setup.VOLUME_LIST_FILE, volume_file_content)
 
 
 # TODO: Create volume list file
